refactor: article을  content로 변경 및 contents.ts 적용
