+++ src/components/cardItem/content/banner/bannerAvatar/index.tsx
@@ -1,10 +1,10 @@
 import { Avatar } from '@/components/common';
 import { useState, useEffect } from 'react';
-import { companyProps } from '@/components/cardItem/article';
 import * as style from './style.css';
 import favicon from '@/assets/favicon.ico';
+import { company } from '@/types/contents';
 
-const BannerAvatar = ({ companies }: { companies?: companyProps[] }) => {
+const BannerAvatar = ({ companies }: { companies?: company[] }) => {
   if (!companies) {
     companies = [{ companyName: '아직 사람들이', companyLogoImgUrl: favicon }];
   }
