+++ src/components/cardItem/content/banner/bannerAvatar/style.css.ts
