+++ src/components/cardItem/content/banner/bannerText/style.css.ts
