+++ src/components/cardItem/content/cardBottom/style.css.ts
