+++ src/components/cardItem/content/cardTop/style.css.ts
