+++ src/components/cardItem/content/index.tsx
@@ -2,28 +2,10 @@ import { Card } from '@/components/common';
 import CardTop from './cardTop';
 import CardBottom from './cardBottom';
 import * as style from './style.css';
-
-export type companyProps = {
-  companyName: string;
-  companyLogoImgUrl: string;
-};
-
-export type ArticleCardProps = {
-  contentId?: number;
-  link: string;
-  contentImgUrl: string;
-  isBookmarked: boolean;
-  isLiked: boolean;
-  likeCount: number;
-  viewCount: number;
-  creatorName: string;
-  createdAt: string;
-  title: string;
-  recommendationCompanies?: companyProps[];
-};
+import { content } from '@/types/contents';
 
 // props: 링크, 이미지, 북마크, 하트, 조회수, 크리에이터 이름, 날짜, 제목, 회사, 회사 아바타
-const ArticleCard = ({
+const ContentCard = ({
   link,
   contentImgUrl,
   isBookmarked,
@@ -34,15 +16,15 @@ const ArticleCard = ({
   createdAt,
   title,
   recommendationCompanies,
-}: ArticleCardProps) => {
+}: content) => {
   /*
     TODO
-    1. 아티클 카드 클릭 시 해당 링크로 이동
+    1. 콘텐츠 카드 클릭 시 해당 링크로 이동
     2. 북마크, 하트 아이콘 클릭 시, 북마크, 하트 여부에 따라 북마크, 하트 or 북마크 하트 취소
-    3. Article Card 데이터 API가 오면 props가 card data 1개로 변하니 나중에 수정할 것
+    3. Content Card 데이터 API가 오면 props가 card data 1개로 변하니 나중에 수정할 것
    */
   return (
-    <Card type="article">
+    <Card type="content">
       <article className={style.cardItem}>
         <div className={style.cardContainer}>
           <CardTop
@@ -66,4 +48,4 @@ const ArticleCard = ({
   );
 };
 
-export default ArticleCard;
+export default ContentCard;
