+++ src/components/cardItem/content/style.css.ts
